move top-level jest config to `jest` folder
