+++ packages/browser/jest.config.js
@@ -1,4 +1,4 @@
-const baseConfig = require('../../jest.config.js');
+const baseConfig = require('../../jest/jest.config.js');
 
 module.exports = {
   ...baseConfig,
