+++ packages/core/jest.config.js
@@ -1,1 +1,1 @@
-module.exports = require('../../jest.config.js');
+module.exports = require('../../jest/jest.config.js');
